requirements.txt
# Core dependencies
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Async HTTP client
aiohttp>=3.9.0
requests>=2.31.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Analysis and visualization (for future use)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Utilities
tqdm>=4.65.0

